iot-pipeline/generator/pyproject.toml
[tool.poetry]
name = "iot-sensor-generator"
version = "0.1.0"
description = "IoT Sensor Data Generator for Kafka"
authors = ["hkkim"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
kafka-python-ng = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"